break files to granular level. The next push would comprise of collecting all common wallet code into a library
